Changes in JSBML required changes here.

+++ src/org/sbml/squeezer/io/LaTeXExport.java
@@ -409,7 +409,7 @@ public class LaTeXExport extends LaTeX {
 			buffer = brackets(buffer);
 		if (u.getExponent() != 1) {
 			buffer.append("^{");
-			buffer.append(Integer.toString(u.getExponent()));
+			buffer.append(Double.toString(u.getExponent()));
 			buffer.append('}');
 		}
 		return buffer;
